requirements.txt
PySide6
Pillow
# ffmpeg should be installed on the system
